CCM-5505: Split timeout tests into their own files
